Extract pasteboard code into its own type and impl
